+++ pygsp/filters/expwin.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from __future__ import division
+
 import numpy as np
 
 from . import Filter  # prevent circular import in Python < 3.5
@@ -11,10 +13,14 @@ class Expwin(Filter):
     Parameters
     ----------
     G : graph
-    bmax : float
-        Maximum relative band (default = 0.2)
-    a : int
-        Slope parameter (default = 1)
+    band_min : float
+        Minimum relative band. The filter take the value 0.5 at this relative
+        frequency (default = None).
+    band_max : float
+        Maximum relative band. The filter take the value 0.5 at this relative
+        frequency (default = 0.2).
+    slope : float
+        The slope at cut-off (default = 1).
 
     Examples
     --------
@@ -25,7 +31,7 @@ class Expwin(Filter):
     >>> G = graphs.Ring(N=20)
     >>> G.estimate_lmax()
     >>> G.set_coordinates('line1D')
-    >>> g = filters.Expwin(G)
+    >>> g = filters.Expwin(G, band_min=0.1, band_max=0.7, slope=5)
     >>> s = g.localize(G.N // 2)
     >>> fig, axes = plt.subplots(1, 2)
     >>> g.plot(ax=axes[0])
@@ -33,24 +39,42 @@ class Expwin(Filter):
 
     """
 
-    def __init__(self, G, bmax=0.2, a=1.):
+    def __init__(self, G, band_min=None, band_max=0.2, slope=1):
 
-        def fx(x, a):
-            y = np.exp(-float(a)/x)
-            if isinstance(x, np.ndarray):
-                y = np.where(x <= 0, 0., y)
-            else:
-                if x <= 0:
-                    y = 0.
-            return y
+        self.band_max = band_max
+        self.band_min = band_min
+        self.slope = slope
 
-        def gx(x, a):
-            y = fx(x, a)
-            return y/(y + fx(1 - x, a))
+        def f(x):
+            # Canary to avoid division by zero and overflow.
+            y = np.where(x <= 0, -1, x)
+            y = np.exp(-slope / y)
+            return np.where(x <= 0, 0, y)
 
-        def ffin(x, a):
-            return gx(1 - x, a)
+        def h(x):
+            y = f(x)
+            z = f(1 - x)
+            return y / (y + z)
 
-        g = [lambda x: ffin(np.float64(x)/bmax/G.lmax, a)]
+        def g_highpass(x):
+            return h(0.5 - x/G.lmax + band_max)
+        def g_lowpass(x):
+            return h(0.5 + x/G.lmax - band_min)
+
+        if band_min is None:
+            g = [g_highpass]
+        elif band_max is None:
+            g = [g_lowpass]
+        else:
+            g = [lambda x: g_lowpass(x) * g_highpass(x)]
 
         super(Expwin, self).__init__(G, g)
+
+    def _get_extra_repr(self):
+        attrs = dict()
+        if self.band_min is not None:
+            attrs.update(band_min='{:.2f}'.format(self.band_min))
+        if self.band_max is not None:
+            attrs.update(band_max='{:.2f}'.format(self.band_max))
+        attrs.update(slope='{:.0f}'.format(self.slope))
+        return attrs