+++ pygsp/filters/simpletight.py
@@ -89,13 +89,14 @@ class SimpleTight(Filter):
 
         if not scales:
             scales = (1./(2.*G.lmax) * np.power(2, np.arange(Nf-2, -1, -1)))
+        self.scales = scales
 
         if len(scales) != Nf - 1:
             raise ValueError('len(scales) should be Nf-1.')
 
-        g = [lambda x: kernel(scales[0] * x, 'sf')]
+        kernels = [lambda x: kernel(scales[0] * x, 'sf')]
 
         for i in range(Nf - 1):
-            g.append(lambda x, i=i: kernel(scales[i] * x, 'wavelet'))
+            kernels.append(lambda x, i=i: kernel(scales[i] * x, 'wavelet'))
 
-        super(SimpleTight, self).__init__(G, g)
+        super(SimpleTight, self).__init__(G, kernels)