+++ pygsp/graphs/randomring.py
@@ -31,6 +31,8 @@ class RandomRing(Graph):
 
     def __init__(self, N=64, seed=None, **kwargs):
 
+        self.seed = seed
+
         rs = np.random.RandomState(seed)
         position = np.sort(rs.uniform(size=N), axis=0)
 
@@ -49,6 +51,8 @@ class RandomRing(Graph):
         coords = np.stack([np.cos(angle), np.sin(angle)], axis=1)
         plotting = {'limits': np.array([-1, 1, -1, 1])}
 
-        super(RandomRing, self).__init__(W=W, gtype='random-ring',
-                                         coords=coords, plotting=plotting,
+        super(RandomRing, self).__init__(W=W, coords=coords, plotting=plotting,
                                          **kwargs)
+
+    def _get_extra_repr(self):
+        return dict(seed=self.seed)